Refactor Sparsemax implementation: move SparsemaxFunction to sparsemax_func_v2 and update imports
