+++ activations_plus/sparsemax/__init__.py
@@ -1,6 +1,6 @@
 """Sparsemax activation package initialization."""
 
-from .sparsemax_func import SparsemaxFunction
+from .sparsemax_func_v2 import SparsemaxFunction
 from .sparsemax_v2 import Sparsemax
 from .utils import flatten_all_but_nth_dim, unflatten_all_but_nth_dim
